+++ web-wallet/CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add "Support" section under Settings [#3071]
 - Add user feedback for "Send" flow validation [#3098]
+- Add validation for self-referential transactions [#3099]
 
 ### Changed
 
@@ -407,6 +408,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#3076]: https://github.com/dusk-network/rusk/issues/3076
 [#3081]: https://github.com/dusk-network/rusk/issues/3081
 [#3098]: https://github.com/dusk-network/rusk/issues/3098
+[#3099]: https://github.com/dusk-network/rusk/issues/3099
 
 <!-- VERSIONS -->